+++ src/window.rs
@@ -750,6 +750,17 @@ impl Window {
         self.window.set_decorations(decorations)
     }
 
+    /// Gets the window's current decorations state.
+    ///
+    /// ## Platform-specific
+    ///
+    /// - **Wayland / X11:** Not implemented.
+    /// - **iOS / Android / Web:** Unsupported.
+    #[inline]
+    pub fn is_decorated(&self) -> bool {
+        self.window.is_decorated()
+    }
+
     /// Change whether or not the window will always be on top of other windows.
     ///
     /// ## Platform-specific